reverted the platoform ui modifications in components
