--- extensions/default/src/Components/ProgressLoadingBar/index.js
@@ -1,2 +1,0 @@
-import ProgressLoadingBar from './ProgressLoadingBar';
-export default ProgressLoadingBar;
