--- extensions/default/src/customizations/contextMenuItemCustomization.ts
@@ -1,5 +1,0 @@
-import ContextMenuItemComponent from '../Components/ContextMenuItemComponent';
-
-export default {
-  'ui.ContextMenuItem': ContextMenuItemComponent,
-};
