+++ extensions/default/src/customizations/loadingIndicatorProgressCustomization.tsx
@@ -1,4 +1,4 @@
-import LoadingIndicatorProgress from '../Components/LoadingIndicatorProgress';
+import { LoadingIndicatorProgress } from '@ohif/ui';
 
 export default {
   'ui.LoadingIndicatorProgress': LoadingIndicatorProgress,
