+++ extensions/default/src/customizations/progressLoadingBarCustomization.tsx
@@ -1,4 +1,4 @@
-import ProgressLoadingBar from '../Components/ProgressLoadingBar';
+import { ProgressLoadingBar } from '@ohif/ui';
 
 export default {
   'ui.ProgressLoadingBar': ProgressLoadingBar,
