+++ extensions/default/src/customizations/viewportActionCornersCustomization.ts
@@ -1,4 +1,4 @@
-import ViewportActionCorners from '../Components/ViewportActionCorners';
+import { ViewportActionCorners } from '@ohif/ui';
 
 export default {
   'ui.ViewportActionCorner': ViewportActionCorners,
