+++ extensions/default/src/getCustomizationModule.tsx
@@ -14,7 +14,6 @@ import onDropHandlerCustomization from './customizations/onDropHandlerCustomizat
 import loadingIndicatorProgressCustomization from './customizations/loadingIndicatorProgressCustomization';
 import loadingIndicatorTotalPercentCustomization from './customizations/loadingIndicatorTotalPercentCustomization';
 import progressLoadingBarCustomization from './customizations/progressLoadingBarCustomization';
-import contextMenuItemCustomization from './customizations/contextMenuItemCustomization';
 import viewportActionCornersCustomization from './customizations/viewportActionCornersCustomization';
 import labellingFlowCustomization from './customizations/labellingFlowCustomization';
 
@@ -57,7 +56,6 @@ export default function getCustomizationModule({ servicesManager, extensionManag
         ...loadingIndicatorProgressCustomization,
         ...loadingIndicatorTotalPercentCustomization,
         ...progressLoadingBarCustomization,
-        ...contextMenuItemCustomization,
         ...viewportActionCornersCustomization,
         ...labellingFlowCustomization,
       },
