+++ platform/ui/src/components/LoadingIndicatorTotalPercent/LoadingIndicatorTotalPercent.tsx
@@ -1,6 +1,8 @@
-import CustomizableRenderComponent from '../../utils/CustomizableRenderComponent';
+import React from 'react';
 
-export interface LoadingIndicatorTotalPercentProps {
+import LoadingIndicatorProgress from '../LoadingIndicatorProgress';
+
+interface Props {
   className?: string;
   totalNumbers: number | null;
   percentComplete: number | null;
@@ -16,17 +18,32 @@ function LoadingIndicatorTotalPercent({
   className,
   totalNumbers,
   percentComplete,
-  loadingText,
-  targetText,
-}: LoadingIndicatorTotalPercentProps) {
-  return CustomizableRenderComponent({
-    customizationId: 'ui.LoadingIndicatorTotalPercent',
-    className,
-    totalNumbers,
-    percentComplete,
-    loadingText,
-    targetText,
-  });
+  loadingText = 'Loading...',
+  targetText = 'segments',
+}: Props): JSX.Element {
+  const progress = percentComplete;
+  const totalNumbersText = totalNumbers !== null ? `${totalNumbers}` : '';
+  const numTargetsLoadedText =
+    percentComplete !== null ? Math.floor((percentComplete * totalNumbers) / 100) : '';
+
+  const textBlock =
+    !totalNumbers && percentComplete === null ? (
+      <div className="text-sm text-white">{loadingText}</div>
+    ) : !totalNumbers && percentComplete !== null ? (
+      <div className="text-sm text-white">Loaded {percentComplete}%</div>
+    ) : (
+      <div className="text-sm text-white">
+        Loaded {numTargetsLoadedText} of {totalNumbersText} {targetText}
+      </div>
+    );
+
+  return (
+    <LoadingIndicatorProgress
+      className={className}
+      progress={progress}
+      textBlock={textBlock}
+    />
+  );
 }
 
 export default LoadingIndicatorTotalPercent;
