+++ platform/ui/src/components/ProgressLoadingBar/ProgressLoadingBar.tsx
@@ -1,4 +1,6 @@
-import CustomizableRenderComponent from '../../utils/CustomizableRenderComponent';
+import React, { ReactElement } from 'react';
+
+import './ProgressLoadingBar.css';
 
 export type ProgressLoadingBarProps = {
   progress?: number;
@@ -9,12 +11,22 @@ export type ProgressLoadingBarProps = {
  * If progress is provided, it will render a progress bar
  * The progress text can be optionally displayed to the left of the bar.
  */
-
-function ProgressLoadingBar({ progress }) {
-  return CustomizableRenderComponent({
-    customizationId: 'ui.ProgressLoadingBar',
-    progress,
-  });
+function ProgressLoadingBar({ progress }: ProgressLoadingBarProps): ReactElement {
+  return (
+    <div className="loading">
+      {progress === undefined || progress === null ? (
+        <div className="infinite-loading-bar bg-primary-light"></div>
+      ) : (
+        <div
+          className="bg-primary-light"
+          style={{
+            width: `${progress}%`,
+            height: '8px',
+          }}
+        ></div>
+      )}
+    </div>
+  );
 }
 
 export default ProgressLoadingBar;
